split color toolkit and types
