+++ src/components/color/types.ts
@@ -5,22 +5,28 @@ export type ColorProps = {
   raw: AnyColor;
 };
 
+export type ColorActionProps = {
+  action: Dispatch<SetStateAction<AnyColor>>;
+};
+
+export type ColorHexProps = {
+  raw: string;
+} & ColorActionProps;
+
+export type ColorHslProps = {
+  raw: HslaColor;
+} & ColorActionProps;
+
+export type ColorRgbProps = {
+  raw: RgbaColor;
+} & ColorActionProps;
+
+export type ColorPickerProps = {
+  raw: HsvaColor;
+} & ColorActionProps;
+
 export type ColorInputProps = {
   update?: ChangeEventHandler<HTMLInputElement>;
   focus?: ChangeEventHandler<HTMLInputElement>;
   leave?: ChangeEventHandler<HTMLInputElement>;
 } & ComponentPropsWithoutRef<"input">;
-
-export type ColorPickerProps = {
-  raw: HsvaColor;
-  action: Dispatch<SetStateAction<AnyColor>>;
-};
-
-export type ColorToolkitProps = {
-  color: {
-    hex: string;
-    hsl: HslaColor;
-    rgb: RgbaColor;
-  };
-  action: Dispatch<SetStateAction<AnyColor>>;
-};
